Move dnd-parsing logic to /dnd
