+++ src/dnd/spells.ts
@@ -9,8 +9,8 @@ import {
     parseRange,
     parseSpellLevel,
     parseSpellSchool,
-} from './parser';
-import { getImageUrl, getSpellsUrl } from './urls';
+} from '../parser';
+import { getImageUrl, getSpellsUrl } from '../urls';
 
 interface Caster {
     name: string;
