+++ src/main.ts
@@ -1,23 +1,23 @@
-import { write, writeFileSync } from 'fs';
-import { getConditionsStatusesAndDiseases } from './conditions';
+import { writeFileSync } from 'fs';
+import { getConditionsStatusesAndDiseases } from './dnd/conditions';
 import { loadData } from './data';
-import { getSpells } from './spells';
-import { getCreatures } from './creatures';
+import { getSpells } from './dnd/spells';
+import { getCreatures } from './dnd/creatures';
 import { StopwatchLogger } from './util';
-import { getClassesAndClassFeats } from './classes';
-import { getItems } from './items';
-import { getRules } from './rules';
-import { getActions } from './actions';
-import { getFeats } from './feats';
-import { getLanguages } from './languages';
-import { getNames } from './names';
-import { getBackgrounds } from './backgrounds';
-import { getTables } from './tables';
-import { getSpecies } from './species';
-import { getSources } from './sources';
-import { getTrapsAndHazards } from './hazards';
-import { getObjects } from './objects';
-import { getVehicles } from './vehicles';
+import { getClassesAndClassFeats } from './dnd/classes';
+import { getItems } from './dnd/items';
+import { getRules } from './dnd/rules';
+import { getActions } from './dnd/actions';
+import { getFeats } from './dnd/feats';
+import { getLanguages } from './dnd/languages';
+import { getNames } from './dnd/names';
+import { getBackgrounds } from './dnd/backgrounds';
+import { getTables } from './dnd/tables';
+import { getSpecies } from './dnd/species';
+import { getSources } from './dnd/sources';
+import { getTrapsAndHazards } from './dnd/hazards';
+import { getObjects } from './dnd/objects';
+import { getVehicles } from './dnd/vehicles';
 
 function main(): void {
     const stopwatch = new StopwatchLogger();
