+++ tsconfig.json
@@ -51,9 +51,7 @@
         // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
         // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
     },
-    "include": [
-        "src/*"
-    ],
+    "include": ["src/*", "src/dnd/*"],
     "exclude": [
         "tests/*"
     ]
